+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ---
-## [X.X.X]
+## [1.3.3]
 * Feature:
 	* Seek media with a given offset
 * Demo: